==> 01 data-import-and-exploration.ipynb <==
from glob import glob
import pandas as pd
import numpy as np

# how do these files work

filenames = glob("GYAFC_Corpus\\Entertainment_Music\\test\\*")

filenames

formal = open(filenames[0], encoding='utf-8')
formal = formal.readlines()

formal_0 = open(filenames[1], encoding='utf-8')
formal_0 = formal_0.readlines()

formal_1 = open(filenames[2], encoding='utf-8')
formal_1 = formal_1.readlines()

formal_2 = open(filenames[3], encoding='utf-8')
formal_2 = formal_2.readlines()

formal_3 = open(filenames[4], encoding='utf-8')
formal_3 = formal_3.readlines()

informal = open(filenames[5], encoding='utf-8')
informal = informal.readlines()

informal_0 = open(filenames[6], encoding='utf-8')
informal_0 = informal_0.readlines()

informal_1 = open(filenames[7], encoding='utf-8')
informal_1 = informal_1.readlines()

informal_2 = open(filenames[8], encoding='utf-8')
informal_2 = informal_2.readlines()

informal_3 = open(filenames[9], encoding='utf-8')
informal_3 = informal_3.readlines()

x = 7

print(formal[x])
print(informal_0[x])
print(informal_1[x])
print(informal_2[x])
print(informal_3[x])

formal[0][:-2]

# put in dataframe for easier use later

def files_to_df():
    # takes all files and puts data into dataframes 
    # (one for formal -> informal, one for informal -> formal)
    # "tune" and "test" datasets give 3 reference outputs for each input.
    # "train" dataset only has 1 (Target 0)
    
    f2i_file_ref = {'formal' : 'Original',
               'informal.ref0' : 'Target 0',
               'informal.ref1' : 'Target 1',
               'informal.ref2' : 'Target 2',
               'informal.ref3' : 'Target 3'}
    i2f_file_ref = {'informal' : 'Original',
               'formal.ref0' : 'Target 0',
               'formal.ref1' : 'Target 1',
               'formal.ref2' : 'Target 2',
               'formal.ref3' : 'Target 3'}
    
    f2i_df = pd.DataFrame(columns=['Original',
                                  'Target 0',
                                  'Target 1',
                                  'Target 2',
                                  'Target 3',
                                  'Category',
                                  'Dataset'])
    i2f_df = pd.DataFrame(columns=['Original',
                                    'Target 0',
                                    'Target 1',
                                    'Target 2',
                                    'Target 3',
                                    'Category',
                                    'Dataset'])
    
    categories = ["Entertainment_Music", "Family_Relationships"]
    datasets = ["train", "tune", "test"]
    
    for cat in categories:
        for ds in datasets:
            
            f2i_df_temp = pd.DataFrame(columns = f2i_df.columns.values)
            i2f_df_temp = pd.DataFrame(columns = i2f_df.columns.values)
            
            filenames = glob("GYAFC_Corpus\\" + cat + "\\" + ds + "\\*")
            
            for fn in filenames:
                name = fn[fn.rfind('\\')+1:] # formal, informal, formal.ref0, informal.ref3, etc
                f = open(fn, encoding="utf-8")
                data = [sent for sent in f.readlines()]
                
                if name in i2f_file_ref:
                    i2f_df_temp[i2f_file_ref[name]] = data
                if name in f2i_file_ref:
                    f2i_df_temp[f2i_file_ref[name]] = data
                if ds == 'train':
                    if name == 'formal':
                        i2f_df_temp['Target 0'] = data
                    elif name == 'informal':
                        f2i_df_temp['Target 0'] = data
                    
            f2i_df_temp['Category'] = [cat] * f2i_df_temp.shape[0]
            i2f_df_temp['Category'] = [cat] * i2f_df_temp.shape[0]
            f2i_df_temp['Dataset'] = [ds] * f2i_df_temp.shape[0]
            i2f_df_temp['Dataset'] = [ds] * i2f_df_temp.shape[0]
                
            f2i_df = pd.concat([f2i_df_temp, f2i_df], axis = 0).reset_index(drop=True)
            i2f_df = pd.concat([i2f_df_temp, i2f_df], axis = 0).reset_index(drop=True)
    
    return f2i_df, i2f_df

formal_to_informal_df, informal_to_formal_df = files_to_df()

formal_to_informal_df.to_pickle('formal_to_informal_df.pkl')

informal_to_formal_df.to_pickle('informal_to_formal_df.pkl')
