==> 02 rule-based-approach.ipynb <==
import pandas as pd
import re
from nltk import word_tokenize
import nltk
import numpy as np
import random

# import data

informal_to_formal_df = pd.read_pickle('data/informal_to_formal_df.pkl')
formal_to_informal_df = pd.read_pickle('data/formal_to_informal_df.pkl')

formal_to_informal_df[formal_to_informal_df['Dataset'] == 'train'].head()

# implement rule-based approach

# Notes:
# Nem tudom, miért mondja Uncle Sam, hogy én tettem. <- why
# Vseki den sym piian, vseki den - sram! <- why???

### Covered in their rule-based system
# Capitalization (51%)
## Capitalize 1st letter
## Capitalize letter I
## Proper nouns (NNP/NNPS)
# Punctuation (69%)
# Contractions (8%)
## Handcrafted contractions list (expand)
# Lowercase (8%)
## Lowercase sentences with only uppercase letters
# Normalization (61%)
## Slang
## Censor swear words
# Repetitions (5%)
## Remove multiple characters/repetition (??? !!! reallllly)

### NOT covered in rule-based system
# Delete fillers ("Well")
# Completion (ellipsis)
# Spelling (should do this)
# Split Sentences

informal = formal_to_informal_df['Target 0']
formal = formal_to_informal_df['Original']

for i in range(10):
    n = random.randint(1, len(informal))
    print(informal[n] + correct(informal[n])+ '\n' + formal[n])

text = "i like HBO"
print(text.capitalize())

ex = "Go to a-zlyrics.com don't worry.I am shure..."
correct(ex)

def normalize(sentence):
    # "chatspeak" to "normal" text
    sentence = re.sub(r'\bim\b', "I'm", sentence, flags=re.I) #re.I = ignore case when matching
    sentence = re.sub(r"\bi'm\b", "I'm", sentence)
    sentence = re.sub(r"\bive\b", "I've", sentence, flags=re.I)
    sentence = re.sub(r"\bi've\b", "I've", sentence)
    sentence = re.sub(r"\bill\b", "I'll", sentence, flags=re.I)
    sentence = re.sub(r"\bi'll\b", "I'll", sentence)
    sentence = re.sub(r"\bi\b", "I", sentence)
    sentence = re.sub(r"\bId\b", "I'd", sentence)
    sentence = re.sub(r"\br\b", "are", sentence, flags=re.I)
    sentence = re.sub(r"\bu\b", "you", sentence, flags=re.I)
    sentence = re.sub(r"\bur\b", "your", sentence, flags=re.I)
    sentence = re.sub(r"\burself\b", "yourself", sentence, flags=re.I)
    sentence = re.sub(r"\byoure\b", "you're", sentence, flags=re.I)
    sentence = re.sub(r"\buve\b", "you've", sentence, flags=re.I)
    sentence = re.sub(r"\bu've\b", "you've", sentence, flags=re.I)
    sentence = re.sub(r"\bull\b", "you'll", sentence, flags=re.I)
    sentence = re.sub(r"\bu'll\b", "you'll", sentence, flags=re.I)
    sentence = re.sub(r"\balot\b", "a lot", sentence, flags=re.I)
    sentence = re.sub(r"\bcant\b", "can't", sentence, flags=re.I)
    sentence = re.sub(r"\bisnt\b", "isn't", sentence, flags=re.I)
    sentence = re.sub(r"\bdont\b", "don't", sentence, flags=re.I)
    sentence = re.sub(r"\bthats\b", "that's", sentence, flags=re.I)
    sentence = re.sub(r"\btheyll\b", "they'll", sentence, flags=re.I)
    sentence = re.sub(r"\btheyre\b", "they're", sentence, flags=re.I)
    sentence = re.sub(r"\bhes\b", "he's", sentence, flags=re.I)
    sentence = re.sub(r"\bshes\b", "she's", sentence, flags=re.I)
    sentence = re.sub(r"\bshud\b", "should", sentence, flags=re.I)
    sentence = re.sub(r"\bshld\b", "should", sentence, flags=re.I)
    sentence = re.sub(r"\bcya\b", "see you", sentence, flags=re.I)
    sentence = re.sub(r"\bluv\b", "love", sentence, flags=re.I)
    sentence = re.sub(r"\bbtw\b", "by the way", sentence, flags=re.I)
    sentence = re.sub(r"\bb/c ", "because ", sentence, flags=re.I)
    sentence = re.sub(r"\bw/ ", "with ", sentence, flags=re.I)
    sentence = re.sub(r"\bw/o ", "without ", sentence, flags=re.I)
    sentence = re.sub(r"\bluv\b", "love", sentence, flags=re.I)
    sentence = re.sub(r"\bhavent\b", "haven't", sentence, flags=re.I)
    sentence = re.sub(r"\btuff\b", "tough", sentence, flags=re.I)
    sentence = re.sub(r"\bgurl\b", "girl", sentence, flags=re.I)
    sentence = re.sub(r"\bdoesnt\b", "doesn't", sentence, flags=re.I)
    sentence = re.sub(r"\blets\b", "let's", sentence, flags=re.I)
    sentence = re.sub(r"\btheres\b", "there's", sentence, flags=re.I)
    sentence = re.sub(r"\bwhats\b", "what's", sentence, flags=re.I)
    sentence = re.sub(r"\bwhut\b", "what", sentence, flags=re.I)
    sentence = re.sub(r"\balot\b", "a lot", sentence, flags=re.I)
    sentence = re.sub(r"\bcos\b", "because", sentence, flags=re.I)
    sentence = re.sub(r"\bcuz\b", "because", sentence, flags=re.I)
    sentence = re.sub(r"\bbcuz\b", "because", sentence, flags=re.I)
    sentence = re.sub(r"\bteh\b", "the", sentence, flags=re.I)
    sentence = re.sub(r"\bkewl\b", "cool", sentence, flags=re.I)
    sentence = re.sub(r"\bdat\b", "that", sentence, flags=re.I)
    sentence = re.sub(r"\bda\b", "the", sentence, flags=re.I)
    sentence = re.sub(r"\brite\b", "right", sentence, flags=re.I)
    sentence = re.sub(r"\bwont\b", "won't", sentence, flags=re.I)
    sentence = re.sub(r"\bryt\b", "right", sentence, flags=re.I)
    sentence = re.sub(r"\bcud\b", "could", sentence, flags=re.I)
    sentence = re.sub(r"\bcood\b", "could", sentence, flags=re.I)
    sentence = re.sub(r"\byur\b", "your", sentence, flags=re.I)
    sentence = re.sub(r"\bnd\b", "and", sentence, flags=re.I)
    sentence = re.sub(r"\bbout\b", "about", sentence, flags=re.I)
    sentence = re.sub(r"\bwat\b", "what", sentence, flags=re.I)
    sentence = re.sub(r"\bwht\b", "what", sentence, flags=re.I)
    sentence = re.sub(r"\bnite\b", "night", sentence, flags=re.I)
    sentence = re.sub(r"\bwanna\b", "want to", sentence, flags=re.I)
    sentence = re.sub(r"\bjuz\b", "just", sentence, flags=re.I)
    sentence = re.sub(r"\bb4\b", "before", sentence, flags=re.I)
    sentence = re.sub(r"\bdats\b", "that's", sentence, flags=re.I)
    sentence = re.sub(r"\bdey\b", "they", sentence, flags=re.I)
    sentence = re.sub(r"\bdeyr\b", "they're", sentence, flags=re.I)
    sentence = re.sub(r"\bcoz\b", "because", sentence, flags=re.I)
    
    if ' id is ' in sentence:
        sentence = sentence.replace(' id ', " ID ")
    else:
        sentence = sentence.replace(' id ', " I'd ")
    return sentence

def capital(sentence):
    # check if entire sentence is capitalized: if so lowercase all but 1st character
    if sentence == sentence.upper():
        sentence = sentence.capitalize()
    else:
        sentence = sentence[0].upper() + sentence[1:] # capitalize 1st letter
        # capitalize words with POS tag of NNP or NNPs
        # this doesn't seem to do anything for the moment, because NLTK mostly
        # only tags proper nouns which are ALREADY capitalized
#         text = word_tokenize(sentence)
#         tags = nltk.pos_tag(text)
#         for idx in range(len(text)):
#             if tags[idx][1] in ['NNP', 'NNPS']:
#                 sentence = sentence.replace(tags[idx][0],
#                                  tags[idx][0][0].upper() + tags[idx][0][1:])
#                 print(tags[idx][1] + ' ' + tags[idx][0])
    return sentence

def punctuate(sentence):
    # too much, none at all, spacing and punctuation issues
    sentence = re.sub(r'[\?\!]+(?=[\?\!])', '', sentence) # reduce extra punctuation marks
    sentence = re.sub(r'[\.]+[\.]', '... ', sentence) # turn more than 1 period into 3
    sentence = re.sub(r'\s+', ' ', sentence).strip() # remove extra whitespace
    sentence = re.sub(r'(?<=[\.,])(?=[^\s\.])', ' ', sentence) # handle comma and period with no following space
    sentence = re.sub(r"\. com", ".com", sentence) # fix URLs changed by previous command
    sentence = re.sub(r" , ", ", ", sentence)
    sentence = re.sub(r" \. ", ". ", sentence)
    return sentence

def repeat(sentence):
    # i'd find out what realllllllllllllly happened -> I would determine what really happened
    sentence = re.sub(r'([a-zA-Z])\1{2}', "\1", sentence)
    return sentence

#def lowercase(sentence):
    # all caps words to lowercase
    #return sentence

#def spellcheck(sentence):
    #return sentence

def correct(sentence):
    if type(sentence) == float:
        if np.isnan(sentence):
            return sentence
    sentence = capital(sentence)
    #sentence = lowercase(sentence)
    sentence = normalize(sentence)
    sentence = repeat(sentence)
    sentence = punctuate(sentence)
    #sentence = spellcheck(sentence)
    return sentence

correct(np.nan)

corrected_df = formal_to_informal_df.copy()

corrected_df['Target 0'] =corrected_df['Target 0'].apply(correct)
corrected_df['Target 1'] =corrected_df['Target 1'].apply(correct)
corrected_df['Target 2'] =corrected_df['Target 2'].apply(correct)
corrected_df['Target 3'] =corrected_df['Target 3'].apply(correct)

def strip_whitespace(sentence):
    sentence = re.sub(r'\s+', ' ', sentence).strip()
    return sentence

corrected_df['Original'] = corrected_df['Original'].apply(strip_whitespace)

corrected_df

corrected_df.to_pickle('data/rule_based_corrected_df.pkl')
