==> 03 train-word-embeddings.ipynb <==
import numpy as np
from nltk.tokenize import sent_tokenize, word_tokenize
import pandas as pd
import pickle
import xml.etree.ElementTree as ET
import re
from tqdm import tqdm_notebook as tqdm
from gensim.test.utils import datapath, get_tmpfile
from gensim.models import KeyedVectors
from gensim.scripts.glove2word2vec import glove2word2vec

In replicating Rao & Tetreault (2018) we need to vectorize text data using word embeddings (GloVe) trained on the whole Yahoo Answers corpus. Word vector dimensions are 300 and total vocabulary size is 50,000.

Note: "At test time, we replace unknown tokens with the source token that has the highest attention weight."

# prepare data: extract answers from XML file

%cd C:\Users\cramerus\Documents\Thesis\Yahoo Answers Corpus
%ls

str.replace('<br />', '<br />', '')

file = 'FullOct2007.xml'
small_file = 'small_sample.xml'
new_output = 'YahooCorpus.txt'
sample_output = 'SampleCorpus.txt'

def xml_to_txt(input_file, output_file):
    with open(output_file, 'w', encoding='utf-8') as f:
        for event, elem in tqdm(ET.iterparse(input_file, events=('start', 'end'))):
            if event == 'start' and elem.tag == 'vespaadd': # start new question
                texts = []
                language = None
            if elem.tag == 'answer_item':
                if elem.text == None:
                    continue
                text = str.replace(elem.text, '<br />', '') # remove tags
                text = re.sub(r'\s+', ' ', text).strip() # remove whitespace
                if text not in texts:
                    texts.append(text)
            if elem.tag == 'language':
                language = elem.text # should be None or en-us

            if event == 'end' and elem.tag == 'vespaadd':
                if language == None or language == 'en-us':
                    for text in texts:
                        tokens = word_tokenize(text)
                        for token in tokens:
                            f.write(token + ' ')
                        f.write('\n')
                elem.clear()

xml_to_txt(small_file, sample_output)

xml_to_txt(file, new_output)

# train glove model on corpus, test

# done in command line
# input is Yahoo Answers corpus, created in the correct format above

# $ cd ./Documents
# $ GloVe/build/vocab_count -max-vocab 50000 -verbose 2 < YahooAnswers/YahooCorpus.txt > YahooAnswers/YahooVocab.txt
# $ GloVe/build/cooccur -memory 10.0 -vocab-file YahooAnswers/YahooVocab.txt -verbose 2 -window-size 15 < YahooAnswers/YahooCorpus.txt > YahooAnswers/YahooCooccur.bin
# $ GloVe/build/shuffle -memory 10.0 -verbose 2 < YahooAnswers/YahooCooccur.bin > YahooAnswers/YahooCooccur.shuf.bin
# $ GloVe/build/glove -save-file YahooAnswers/YahooVectors -threads 8 -input-file YahooAnswers/YahooCooccur.shuf.bin -vector-size 300 -binary 2 -vocab-file YahooAnswers/YahooVocab.txt -verbose 2

# output is at Documents/YahooAnswers/YahooVectors.txt

% cd /home/rebekah/Documents/YahooAnswers 
glove = 'YahooVectors.txt'
w2v = "YahooVectors.w2v.txt"
glove2word2vec(glove, w2v)

model = KeyedVectors.load_word2vec_format("YahooVectors.w2v.txt")

model.most_similar('omg')
