==> 04 nmt baseline.ipynb <==
import numpy as np
from nltk.tokenize import word_tokenize
import pandas as pd
import pickle
import tqdm
import torch
from nltk import word_tokenize, pos_tag, ne_chunk

# prepare text files to use with OpenNMT

all_df = pd.read_pickle('data/rule_based_corrected_df.pkl')
all_df.head()

train_src = all_df[all_df['Dataset'] == 'train']['Original']
train_tgt = all_df[all_df['Dataset'] == 'train']['Target 0']
val_src = all_df[all_df['Dataset'] == 'tune']['Original']
val_tgt = all_df[all_df['Dataset'] == 'tune']['Target 0']

assert len(train_src) == len(train_tgt)
assert len(val_src) == len(val_tgt)

def series_to_txt(series, filename):
    with open(filename, 'w', encoding='utf-8') as f:
        for item in series:
            tokens = word_tokenize(item)
            tokens = ' '.join(tokens)
            f.write(tokens + '\n')

series_to_txt(val_src, 'data/src-val.txt')
series_to_txt(val_tgt, 'data/tgt-val.txt')
series_to_txt(train_src, 'data/src-train.txt')
series_to_txt(train_tgt, 'data/tgt-train.txt')

test_src = all_df[all_df['Dataset'] == 'test']['Original']
test_tgt = all_df[all_df['Dataset'] == 'test']['Target 0']

assert len(test_src) == len(test_tgt)

series_to_txt(test_src, 'data/src-test.txt')
series_to_txt(test_tgt, 'data/tgt-test.txt')

# openNMT with default parameters

torch.cuda.is_available()

# using OpenNMT

# http://opennmt.net/OpenNMT-py/quickstart.html

# % OpenNMT-py/preprocess.py -train_src make-it-sound-less-formal/data/src-train.txt -train_tgt make-it-sound-less-formal/data/tgt-train.txt -valid_src make-it-sound-less-formal/data/src-val.txt -valid_tgt make-it-sound-less-formal/data/tgt-val.txt -save_data make-it-sound-less-formal/data/OpenNMTData

# % tools/embeddings_to_torch.py -emb_file_both "data/YahooCorpus/YahooVectors.txt" -dict_file "data/YahooCorpus/OpenNMTData.vocab.pt" -output_file "data/YahooCorpus/embeddings"

# % python3 train.py -save_model data/YahooCorpus/model -data data/YahooCorpus/OpenNMTData -word_vec_size 300 -pre_word_vecs_enc "data/YahooCorpus/embeddings.enc.pt" -pre_word_vecs_dec "data/YahooCorpus/embeddings.dec.pt" 

# preprocess.py -train_src data/YahooCorpus/src-train.txt -train_tgt data/YahooCorpus/tgt-train.txt -valid_src data/YahooCorpus/src-val.txt -valid_tgt data/YahooCorpus/tgt-val.txt -save_data data/model_pretrained_embed_with_copy/data -dynamic_dict
# python3 train.py -save_model data/model_pretrained_embed_with_copy/ -data data/model_pretrained_embed_with_copy/data -word_vec_size 300 -pre_word_vecs_enc "data/glove_pretrained/embeddings.enc.pt" -pre_word_vecs_dec "data/glove_pretrained/embeddings.dec.pt" -copy_attn
# take 2

# OpenNMT-py/preprocess.py -train_src make-it-sound-less-formal/data/src-train-no-ent.txt -train_tgt make-it-sound-less-formal/data/tgt-train-no-ent.txt -valid_src make-it-sound-less-formal/data/src-val-no-ent.txt -valid_tgt make-it-sound-less-formal/data/tgt-val-no-ent.txt -save_data OpenNMT-py/data/model_pretrained_no_ent_with_copy/data -dynamic_dict
# python3 train.py -save_model data/model_pretrained_no_ent_with_copy/model -data data/model_pretrained_no_ent_with_copy/data -word_vec_size 300 -pre_word_vecs_enc "data/glove_pretrained/embeddings.enc.pt" -pre_word_vecs_dec "data/glove_pretrained/embeddings.dec.pt" -copy_attn


# informal to formal
# OpenNMT-py/preprocess.py -train_src make-it-sound-less-formal/data/OpenNMT\ files/tgt-train-no-ent.txt -train_tgt make-it-sound-less-formal/data/OpenNMT\ files/src-train-no-ent.txt -valid_src make-it-sound-less-formal/data/OpenNMT\ files/tgt-val-no-ent.txt -valid_tgt make-it-sound-less-formal/data/OpenNMT\ files/src-val-no-ent.txt -save_data make-it-sound-less-formal/data/OpenNMT\ files/formal-to-informal-data -dynamic_dict
# python3 OpenNMT-py/train.py -save_model informal-to-formal -data make-it-sound-less-formal/data/OpenNMT\ files/informal-to-formal-data -word_vec_size 300 -pre_word_vecs_enc "embeddings.enc.pt" -pre_word_vecs_dec "embeddings.dec.pt" -copy_attn -world_size 1 -gpu_ranks 0


# Collect source, target, pred in excel sheet

% cd /home/rebekah/Documents/OpenNMT-py/data/YahooCorpus

pred_file = open('pred-test.txt')
pred = [line.rstrip() for line in pred_file.readlines()]
pred_file.close()

source_file = open('src-test.txt')
source = [line.rstrip() for line in source_file.readlines()]
source_file.close()

target_file = open('tgt-test.txt')
target = [line.rstrip() for line in target_file.readlines()]
target_file.close()

v1_results = pd.DataFrame({'Source': source, 'Target': target, 'Prediction': pred})

% cd /home/rebekah/Documents/make-it-sound-less-formal/data/

writer = pd.ExcelWriter('version_1_test.xlsx')
v1_results.to_excel(writer, engine='xlsxwriter')
writer.save()

results_df = pd.read_pickle('data/v1_results.pkl')
results_df.head()

copy_pred_file = open('data/copy-pred-test.txt')
copy_pred = [line.rstrip() for line in copy_pred_file.readlines()]
copy_pred_file.close()

results_df['Copy Prediction'] = copy_pred

copy_no_ent_pred_file = open('data/pred-test-no-ent.txt')
copy_no_ent_pred = [line.rstrip() for line in copy_no_ent_pred_file.readlines()]
copy_no_ent_pred_file.close()

results_df['No Entity Copy Prediction'] = copy_no_ent_pred

results_df.head()

results_df.to_pickle('data/v1_results.pkl')

# Remove NER from dataset for last model type

all_df = pd.read_pickle('data/rule_based_corrected_df.pkl')
all_df.head()

def ents(sent):
    entities = []
    for chunk in nltk.ne_chunk(nltk.pos_tag(nltk.word_tokenize(sent))):
        if hasattr(chunk, 'label'):
            entities.append(' '.join(c[0] for c in chunk))
    return entities

def replace_ents(sent):
    entities = ents(sent)
    for ent in entities:
        sent = sent.replace(ent, 'ENT')
    return sent

sentence = "My aunt just sent me a cd with an amazing tenor named Josh Groban ."
replace_ents(sentence)

tqdm.tqdm.pandas()

all_df.loc[24, 'Original']

all_df['Original'].progress_apply(replace_ents)

all_df['Target 0'].progress_apply(replace_ents)

train_src = all_df[all_df['Dataset'] == 'train']['Original']
train_tgt = all_df[all_df['Dataset'] == 'train']['Target 0']
val_src = all_df[all_df['Dataset'] == 'tune']['Original']
val_tgt = all_df[all_df['Dataset'] == 'tune']['Target 0']

assert len(train_src) == len(train_tgt)
assert len(val_src) == len(val_tgt)

def series_to_txt(series, filename):
    with open(filename, 'w', encoding='utf-8') as f:
        for item in series:
            tokens = word_tokenize(item)
            tokens = ' '.join(tokens)
            f.write(tokens + '\n')
            
series_to_txt(val_src, 'data/src-val-no-ent.txt')
series_to_txt(val_tgt, 'data/tgt-val-no-ent.txt')
series_to_txt(train_src, 'data/src-train-no-ent.txt')
series_to_txt(train_tgt, 'data/tgt-train-no-ent.txt')
