==> 09 collect seed set of formal and informal lexical pairs.ipynb <==
from tqdm import tqdm_notebook as tqdm
import nltk
import numpy as np
import pandas as pd
import json
import re
import os
from collections import Counter, defaultdict
from glob import glob
import gensim
import xml.etree.ElementTree as ET
from ast import literal_eval
from gensim.test.utils import datapath, get_tmpfile
from gensim.models import KeyedVectors
from gensim.scripts.glove2word2vec import glove2word2vec
import sklearn
import pickle
from sklearn.linear_model import LinearRegression

# put existing acrolinx gzt and files into pd df

with open('data/acrolinx_gzt/lf.json') as lfjson:
    lf = json.load(lfjson)

with open('data/acrolinx_gzt/conv-words.json') as cvjson:
    form_with_sugg = json.load(cvjson)

unclean_gzt = []

with open('data/acrolinx_gzt/archaicWords.gzt') as file:
    unclean_gzt.extend(file.readlines())

with open('data/acrolinx_gzt/countFormalPhrases.gzt') as file:
    unclean_gzt.extend(file.readlines())

with open('data/acrolinx_gzt/countLatinExpressions.gzt') as file:
    unclean_gzt.extend(file.readlines())

gzt = {}

# things i noticed and don't want
exceptions = ['use either', '(']

for item in unclean_gzt:
    if item[0] == '@' or item[0] == '#':
        continue
    item = item.strip()
    if len(item) < 1:
        continue
    trigger = False
    for term in exceptions:
        if term in item:
            trigger = True
    if trigger:
        continue
    item = re.sub('\[', '', item)
    item = re.sub('\]', '', item)
    item = re.sub('\n', '', item)
    item = re.sub(';', '', item)
    if '-->' in item:
        pair = [part.strip() for part in item.split('-->')]
        if ',' in pair[0]:
            form_words = [part.strip() for part in pair[0].split(',')]
            for word in form_words:
                gzt[word] = [part.strip() for part in pair[1].split(',')]
        else:
            gzt[pair[0]] = [part.strip() for part in pair[1].split(',')]
    else:
        gzt[item] = np.nan
        
# There are 597 items.

formal = []
informal = []

for word in gzt:
    formal.append(word)
    informal.append(gzt[word])

words = pd.DataFrame()
words['formal'] = formal
words['suggestions'] = informal
words.head()

words.to_pickle('data/acrolinx_gzt/clean_words.pkl')

# add microsoft words to dataframe

with open('data/microsoft/words.pkl', 'rb') as f:
    mic_words = pickle.load(f)

mic_words = list(set([str(x).strip().lower() for x in list(mic_words)]))
mic_words

sugg = [np.nan] * len(mic_words)
sugg += list(words_df['suggestions'])

mic_words += list(words_df['formal'])

mic_df = pd.DataFrame()
mic_df['words'] = mic_words
mic_df['sugg'] = sugg

mic_df.sample(10)

mic_df.to_pickle('data/lexical_repl/all_words_clean.pkl')

# next: extrapolate to the other words using embeddings

words_df = pd.read_pickle('data/lexical_repl/all_words_clean.pkl')

w2v = gensim.models.KeyedVectors.load_word2vec_format('/home/rebekah/Documents/Word Embeddings/GoogleNews-vectors-negative300.bin', binary=True) 

# The basic way.
word = 'oppose'
w2v.most_similar(positive=[word, 'caveat'], negative=['warning'])

def process_word_lists(wl, embed, one_word_only = False):
    if len(wl) == 1:
        if wl[0] in embed:
            return embed[wl[0]]
    elif len(wl) > 1 and one_word_only == False:
        vecs = [0.0] * len(embed['word'])
        for w in wl:
            if w in embed:
                vecs = list(map(sum, zip(vecs, embed[w])))
        if vecs != [0.0] * len(embed['word']):
            return vecs
    return np.nan

def make_data(df, embed, X, y, one_word_only = False):
    X_ph = np.nan * len(df)
    y_ph = np.nan * len(df)
    df[X] = X_ph
    df[X] = df[X].astype(object)
    df[y] = y_ph
    df[y] = df[y].astype(object)
    
    for idx, row in tqdm(words_df.iterrows(), total=len(words_df)):
        formal_words = nltk.word_tokenize(row['words'])
        df.at[idx, X] = process_word_lists(formal_words, embed, one_word_only) 
        informal_words = []
        if type(row['sugg']) != float:
            for word in row['sugg']:
                word = nltk.word_tokenize(word)
                word = process_word_lists(word, embed, one_word_only)
                informal_words.append(word)
        if len(informal_words) > 0:
            df.at[idx, y] = informal_words[0]
                
    return df

words_df = make_data(words_df, w2v, 'X_w2v', 'y_w2v', one_word_only = True)

words_df.dropna(subset=['X_w2v']).sample(10)

# train linear regression

train = words_df.dropna() # only the words that have suggestions and vectors for both

for idx, row in train.iterrows():
    assert len(row['X_w2v']) == 300

X = np.array(list(train['X_w2v']))
y = np.array(list(train['y_w2v']))

lr = LinearRegression().fit(X, y)

lr.coef_.shape

lr.score(X, y)

def predict(word):
    pred = lr.predict([word]).reshape(-1, 1)
    pred = pred.reshape(300,)
    return w2v.similar_by_vector(pred, topn = 10)

np.isnan(np.nan)

# predict for all other words

w2v_pred = []

for idx, row in tqdm(words_df.iterrows(), total=len(words_df)):
    if type(row['X_w2v']) == float:
        w2v_pred.append(None)
        #formal = nltk.word_tokenize(row['words'])
        #vec = process_word_lists(formal, w2v, one_word_only = True)
        #if type(vec) == float:
        #    w2v_pred.append(np.nan)
        #else:
        #    w2v_pred.append(predict(vec))
    else:
        w2v_pred.append(predict(row['X_w2v']))

words_df['pred_w2v'] = w2v_pred

def display(df, num = 10):
    current = 0
    for idx, row in df.sample(frac=1).iterrows():
        if current > num:
            break
        train = type(row['y_w2v']) != float
        if train:
            continue
        print('Original Word:\t' + row['words'])
        #print('Training Data?:\t' + str(train))
        if type(row['sugg']) != float:
            print('Given Answer:\t' + str(row['sugg']))
        else:
            print()
        if type(row['pred_w2v']) != float:
            ans = ''
            for item in row['pred_w2v']:
                ans += item[0] + '\t' + str('%s' % float('%.3g' % item[1])) + '\n\t\t'
            print('Pred Answers:\t' + ans)
        current += 1

display(words_df.dropna(subset=['pred_w2v']))

words_df.to_pickle('data/lexical_repl/all_words_filled.pkl')

with open('data/lexical_repl/doccano_to_check.txt', 'w') as f:
    for idx, row in tqdm(words_df.iterrows(), total=len(words_df)):
        if row['pred_w2v'] != None:
            ans = ''
            for item in row['pred_w2v']:
                ans += item[0] + '\t'
            f.write(row['words'].upper() + '\t' + ans + '\n')

# collect words

with open('data/lexical_repl/doccano_to_check.txt', 'r') as f:
    tests = f.readlines()
print(len(tests))

test_tuples = []

for item in tests:
    split = item.split('\t')[:-1]
    test_tuples.append((split[0], split[1:]))

words = [tup[0] for tup in test_tuples]
print(len(words))
words = list(set(words))
print(len(words))

new_sets = defaultdict(list)

for item in tqdm(words):
    for pair in test_tuples:
        if pair[0] == item:
            for possible in pair[1]:
                new_sets[item].append(possible)
                
for term in new_sets:
    new_sets[term] = list(set([x.lower() for x in new_sets[term]]))

to_be_checked = pd.DataFrame.from_dict(new_sets, 
                                       orient='index')

to_be_checked.to_csv('data/microsoft/to_be_checked.csv')
